Refactor sports and game selection widgets
Moved the SportSelection and GameSelection classes to separate files for better modularity and maintainability. Updated imports across relevant files to reflect these changes, ensuring consistency and separation of concerns in the project structure.
