+++ gui_main.py
@@ -1,15 +1,12 @@
 from PyQt5.QtCore import pyqtSignal, Qt
-from PyQt5.QtWidgets import (QWidget, QHBoxLayout, QVBoxLayout, QButtonGroup,
-                             QScrollArea, QSpacerItem)
+from PyQt5.QtWidgets import QWidget, QHBoxLayout, QVBoxLayout, QSpacerItem
 
-import shared
 from labels.image import Image as Image
 from labels.small_text import SmallText as SmallText
 from buttons.spoiler_toggle import SpoilerToggle as SpoilerToggle
 from buttons.picture_button import PictureButton as PictureButton
-from buttons.radio_sports_button import RadioSportsButton as RadioSportsButton
-from buttons.radio_game_button import RadioGameButton as RadioGameButton
-
+from widgets.game_select import GameSelection as GameSelection
+from widgets.main_sport_select import SportSelection as SportSelection
 
 class MainMenu(QWidget):
 
@@ -70,149 +67,7 @@ class MainMenu(QWidget):
         self.prefs_button_clicked.emit()
 
 
-class SportSelection(QWidget):
-    def __init__(self):
-        super().__init__()
-        self.main_layout = QVBoxLayout()
-        self.setLayout(self.main_layout)
-
-        scroll_area = QScrollArea()
-        scroll_area.setWidgetResizable(True)
-        scroll_area.setMaximumWidth(400)
-        scroll_area.setMaximumHeight(500)
-        scroll_area.setStyleSheet("""  
-            QScrollArea {
-                border: none;
-                background: transparent;
-            }
-            QScrollArea > QWidget > QWidget {
-                background: transparent;
-            }
-            QScrollBar:vertical {
-                border: none;
-                background: transparent;
-                width: 12px;
-                margin: 0px 0px 0px 0px;
-                padding: 0px;
-            }
-            QScrollBar::handle:vertical {
-                background: #888;
-                min-height: 20px;
-                border-radius: 6px;
-            }
-            QScrollBar::add-line:vertical, QScrollBar::sub-line:vertical {
-                border: none;
-                background: none;
-                height: 0px;
-            }
-            QScrollBar::add-page:vertical, QScrollBar::sub-page:vertical {
-                background: none;
-            }
-            QScrollBar::handle:vertical:hover {
-                background: #666;
-            }
-            """)
-        self.main_layout.addWidget(scroll_area)
-
-        container = QWidget(self)
-        self.vbox = QVBoxLayout()
-        container.setLayout(self.vbox)
-
-        scroll_area.setWidget(container)
-
-        self.sports_button_group = QButtonGroup()
-        self.sports_button_group.setExclusive(True)
-        self.update_sports()
-
-    def update_sports(self):
-        while self.vbox.count():
-            item = self.vbox.takeAt(0)
-            widget = item.widget()
-            if widget:
-                self.sports_button_group.removeButton(widget)
-                widget.deleteLater()
-
-        for sports in shared.user_preferences.sports_order.values():
-            radio_button = RadioSportsButton(sports.upper())
-            self.sports_button_group.addButton(radio_button)
-            self.vbox.addWidget(radio_button)
 
 
-class GameSelection(QWidget):
-    def __init__(self):
-        super().__init__()
-        self.main_layout = QHBoxLayout()
-        self.setLayout(self.main_layout)
 
-        scroll_area = QScrollArea()
-        scroll_area.setWidgetResizable(True)
-        scroll_area.setFixedWidth(800)
-        scroll_area.setMaximumHeight(200)
-        scroll_area.setStyleSheet("""
-        QScrollArea {
-                border: none;
-                background: transparent;
-            }
-            QScrollArea > QWidget > QWidget {
-                background: transparent;
-            }
-            QScrollBar:vertical {
-                border: none;
-                background: transparent;
-                padding: 0px;
-            }
-            QScrollBar:horizontal {
-                height: 6px;
-                background: transparent;
-                border-radius: 10px;
-            }
-            QScrollBar::handle:horizontal {
-                background: #888;
-                min-height: 6px;
-                border-radius: 3px;
-            }
-            QScrollBar::add-line:horizontal, QScrollBar::sub-line:horizontal {
-                border: none;
-                background: none;
-                height: 0px;
-            }
-            QScrollBar::add-page:horizontal, QScrollBar::sub-page:horizontal {
-                background: none;
-            }
-            QScrollBar::handle:horizontal:hover {
-                background: #666;
-            }
-        """)
-        self.main_layout.addWidget(scroll_area)
 
-        container = QWidget(self)
-        self.hbox = QHBoxLayout()
-        container.setLayout(self.hbox)
-
-        scroll_area.setWidget(container)
-
-        self.games_button_group = QButtonGroup()
-        self.games_button_group.setExclusive(True)
-        self.update_games()
-
-    def update_games(self):
-        while self.hbox.count():
-            item = self.hbox.takeAt(0)
-            widget = item.widget()
-            if widget:
-                self.games_button_group.removeButton(widget)
-                widget.deleteLater()
-
-        for games in shared.test_games.values():
-            date = games.get("date", None)
-            home = games.get("home", None)
-            home_score = games.get("home_score", None)
-            away = games.get("away", None)
-            away_score = games.get("away_score", None)
-            time = games.get("time", None)
-
-            if time is not None:
-                radio_button = RadioGameButton(date=date, home=home, home_score=home_score,
-                                                       away=away, away_score=away_score, time=time)
-                self.games_button_group.addButton(radio_button)
-                self.hbox.addWidget(radio_button)
